feat: implement modules/cjsToEsm task
